# swagger.json
{"openapi":"3.0.0","info":{"title":"evaluaciones_mongo_crud","description":"API CRUD para la gestion de formatos de evaluación para las entidades de la universidad.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ItemDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"estructura_item":{"type":"object","default":{}},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","estructura_item","fecha_creacion","fecha_modificacion"]},"SeccionDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"estructura_seccion":{"type":"object","default":{}},"items_id":{"default":["_id"],"type":"array","items":{"type":"string"}},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","estructura_seccion","items_id","fecha_creacion","fecha_modificacion"]},"TipoEvaluacionDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"estructura_tipo_evaluacion":{"type":"object","default":{}},"secciones_id":{"default":["_id"],"type":"array","items":{"type":"string"}},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","estructura_tipo_evaluacion","secciones_id","fecha_creacion","fecha_modificacion"]},"EvaluacionDto":{"type":"object","properties":{"tipo_evaluacion_id":{"type":"string","default":"_id"},"nombre":{"type":"string"},"descripcion":{"type":"string"},"estructura_evaluacion":{"type":"object","default":{}},"resultado":{"type":"object","default":{}},"estado":{"type":"object","default":{}},"respuestas_por_fecha":{"type":"object","default":[{"fecha_presentacion":"2020-12-29T02:14:28.746Z","respuestas":[{"item_id":"_id","respuesta":{}}]}]},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["tipo_evaluacion_id","nombre","descripcion","estructura_evaluacion","resultado","estado","respuestas_por_fecha","fecha_creacion","fecha_modificacion"]}}},"paths":{"/items":{"post":{"operationId":"ItemsController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemDto"}}}},"responses":{"201":{"description":""}},"tags":["items"]},"get":{"operationId":"ItemsController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["items"]}},"/items/{id}":{"get":{"operationId":"ItemsController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["items"]},"put":{"operationId":"ItemsController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemDto"}}}},"responses":{"200":{"description":""}},"tags":["items"]},"delete":{"operationId":"ItemsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["items"]}},"/secciones":{"post":{"operationId":"SeccionesController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeccionDto"}}}},"responses":{"201":{"description":""}},"tags":["secciones"]},"get":{"operationId":"SeccionesController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["secciones"]}},"/secciones/{id}":{"get":{"operationId":"SeccionesController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["secciones"]},"put":{"operationId":"SeccionesController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeccionDto"}}}},"responses":{"200":{"description":""}},"tags":["secciones"]},"delete":{"operationId":"SeccionesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["secciones"]}},"/tipos-evaluacion":{"post":{"operationId":"TiposEvaluacionController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TipoEvaluacionDto"}}}},"responses":{"201":{"description":""}},"tags":["tipos_evaluacion"]},"get":{"operationId":"TiposEvaluacionController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tipos_evaluacion"]}},"/tipos-evaluacion/{id}":{"get":{"operationId":"TiposEvaluacionController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tipos_evaluacion"]},"put":{"operationId":"TiposEvaluacionController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TipoEvaluacionDto"}}}},"responses":{"200":{"description":""}},"tags":["tipos_evaluacion"]},"delete":{"operationId":"TiposEvaluacionController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tipos_evaluacion"]}},"/evaluaciones":{"post":{"operationId":"EvaluacionesController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluacionDto"}}}},"responses":{"201":{"description":""}},"tags":["evaluaciones"]},"get":{"operationId":"EvaluacionesController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["evaluaciones"]}},"/evaluaciones/{id}":{"get":{"operationId":"EvaluacionesController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["evaluaciones"]},"put":{"operationId":"EvaluacionesController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluacionDto"}}}},"responses":{"200":{"description":""}},"tags":["evaluaciones"]},"delete":{"operationId":"EvaluacionesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["evaluaciones"]}}}}